requirements.txt
# ── Core API & Server ─────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0

# For production-grade RAG (upgrade from TF-IDF):
# sentence-transformers
scipy
chromadb
# ── LLM Client ────────────────────────────────────────────────────────────────
groq>=0.9.0

# ── Knowledge Graph ───────────────────────────────────────────────────────────
networkx>=3.3

# ── NLP & Fuzzy Matching ──────────────────────────────────────────────────────
spacy>=3.7.0
rapidfuzz>=3.6.0
# After install, download the language model:
# python -m spacy download en_core_web_sm

# ── RAG / Semantic Retrieval ──────────────────────────────────────────────────
sentence-transformers>=2.7.0
scipy>=1.13.0
# chromadb  # optional: uncomment for persistent vector store

# ── Utilities ─────────────────────────────────────────────────────────────────
google-genai>=0.5.0
